docgen.lua: require language server modules
In # 109, when loading nvim_lsp.lua, all the language server lua modules were not loaded,
so the language server module was missing from the configs module referenced in docgen.
So require the language server module in docgen.lua.
configs.lua and utils.lua are also required, but they have no effect as they do not add themselves to the configs module.