Leave nvim_lsp.util.sh buffer open on error
Fixes #196.

If the child process exits with a non-zero exit code or receives a
non-zero signal, leave the buffer with the output open so that the user
has a chance to figure out what went wrong.

The `update_chunk` function is not modified, just moved up a bit, so
that `uv.spawn`'s `on_exit` callback can capture it and use it to
append the exit code and signal to the output buffer.