Use builtin terminal for util.sh output
Addresses #207. Supersedes #201.

Using `termopen()`, `util.sh` gets full terminal functionality for free:

- progress bars leveraging `\r` and other more fancy output types (e.g.
  curses UIs)
- interactive prompts
- canceling with `Ctrl-C`
- etc.

The drawback is that currently (see https://github.com/neovim/neovim/issues/7607),
the `on_exit` callback for `termopen()`/`jobstart()` *has* to be a VimL
function, which means a global VimL function has to be created
on-the-fly with a name which hopefully won't clobber anything that's
already defined.

Like #201, this commit makes sure that the output buffer remains open if
the script exits with anything else than 0.