Remove unnecessary init_options for Metals.
As of the 0.9.0 release of Metals, a bug was fixed where previously all
of the `InitializationOptions` needed to be specified if you set any in
order to not have a null pointer exception. This is fixed now so all of
the default have been removed, and only the necessary ones remain.