Prevent R language server installer from failing
R tries to download the package into a system folder, which
could fail if the user doesn't have write access to it.

This tells R to create a user package directory and install
it there.

From the documentation of `install.packages`:

"If lib is omitted or is of length one and is not a (group) writable
directory, in interactive use the code offers to create a personal
library tree (the first element of Sys.getenv("R_LIBS_USER")) and
install there."