Add gen_cmdline_args as an optional setup function. (#1)
gen_cmdline_args is an optional function that can be provided in the setup table
for a given configuration. When non-nil the function takes the full path of the
active file and will return a table of strings which are appended to the cmd
table prior to launching the LSP server.

By dynamically generating command line arguments users can customize LSP server
configuration based on the project of interest. A more concrete example is that
it allows clangd users to specify the `--compile-commands-dir=` argument when
the `compile_commands.json` file is not directly in the build path for the
project (which is not unusual for CMake based projects) or when multiple
`compile_commands.json` files exist at the same "nested" level.