npm_installer: Skip creating package.json
If the name in the `package.json` file generated by `npm init`
matched the name of one of the packages to be installed, npm would
error out. It turns out that npm will happily install packages without
the current directory containing a `package.json`, so just do that
instead of any gymnastics to avoid name the name collision.

fixes #409