bashls: Set root_dir to nil when editing file in ~
A problem with bash-language-server is that it has a bad habit of trying
to recursively scan the entire home folder if you for example tried to
edit a shell script directly in `~` (e.g. `~/foo.sh`).

This causes a bunch of weird behaviour and causes the server to take
forever to start.

To prevent this we set `root_dir` to `nil` when the resolved `root_dir`
happened to be `~`. This solution is kind of hinted towards if you look
at their own README for how you should configure it with coc.nvim:

    "languageserver": {
      "bash": {
        "command": "bash-language-server",
        "args": ["start"],
        "filetypes": ["sh"],
        "ignoredRootPaths": ["~"]
      }
    }