sumneko: do not send telemetry data by default
By default, sumneko sends telemetry data containing a randomized but unique
identifier to a Chinese IP.

The token can be found in lua-language-server/log/token.

The string sent looks like this:

  $...pulse.XXXXXXXXXXXXXXXX.Neovim 0.5.0.J...platform.XXXXXXXXXXXXXXXX.macOS 64.libc++ libc++ 10000.clang Clang 12.0.0.

In case of errors, these are sent, too. You can visit the IP to get a summary of
the collected data: http://119.45.194.183