efm: Include current file dir in root detection
The current settings for EFM make it so that it will only work within a git repository.

This change adds a fallback, so that if the current file is _not_ inside a git repository
then the current directory is used instead.

This is useful, for example, when opening a random python file, so as to get feedback
from the system flake8.