Fix starting the LSP when creating a new file
As part of looking for the root directory, the code uses
vim.loop.fs_realpath to resolve symlinks. This doesn't work if the path
involved doesn't yet exist (such as when creating a new file). The
current behavior is to give up at this point as the returned 'real' path
is nil, but we can do better by taking the unresolved path and just
allowing the iterate_parents path to get the directory that the file is
in (which will likely exist in most cases). This change does just that -
if fs_realpath returns nil, then we just proceed with the unresolved
path.

This allows language servers to autostart even if the file doesn't yet
exist. Apparently some language servers do require the file to exist
beforehand anyway, but this will allow those without that restriction to
work correctly. I tested it at least with terraform-ls and gopls and it
seems to work fine.