fix(bashls): consider .git when selecting root_dir
When using bashls with plugins like https://github.com/ahmedkhalf/project.nvim
it is very annoying, when the project root suddenly changes, when you
are editing a script, which is a part of a bigger project. This change
should fix the issue.

See also: https://github.com/LunarVim/LunarVim/issues/1467