Update the `on_attach` hook example
Previously the example had the user specifying their LSP servers manually. This change looks at available servers and attaches to all of them. This makes sense if it's more likely that the user will be interested in attaching this hook to all available servers, not just a subset of them.