Fix all server configuration doc links
Several of the links to `server_configurations.md` were broken, as they were missing the `doc/` prefix. Fix all of them, and give them all the same relative URL.

I considered whether to use the absolute URL for all of them, or the relative URL, and it made sense to me to use the relative URL so that a fork could conceivably update the `server_configurations` document and its links would still work. I have no preference there so if you have a preference on what the actual link should be just let me know.