feat(leanls): add root dir as command-line arg
The argument is ignored by the server, but it is useful because it is
shown by `htop`, `ps ax`, etc., so you can figure out which Lean servers
are running.

The lean3ls doesn't require a corresponding change: we use the
`lean-language-server` wrapper, which appends the root directory itself
to the wrapped `lean --server` command.