chore: remove healthcheck (#1636)
Previously, lspconfig had a recommended healthcheck. This was mostly
noise as its predominant use was to check if a server was installed.

This check was not accurate as vim.fn.executable is extremely primitive
and does not consider cmd_env (forwarded to uv_spawn) or language servers
that are run via an interpreter.

Furthermore, checking if a language server is executable will no longer
be a useful signal once TCP support (which allows for connecting to
already running or remote servers).

Remove for now.