chore(elmls): remove overriding elm paths (#1718)
As stated in the [elm-language-server docs](https://github.com/elm-tooling/elm-language-server#server-settings) – the default value of elm binaries should be an empty string so the language server searches for binaries as local project installs and only use global ones as fallback. If we specify it as the standard "global values" like this config is currently doing, we're just going straight to the fallback, breaking compatibility with local installs.