fix(lspstop): make LspStop behave as described in the docs (#1764)
This bug was introduced in #1324.

When calling `:LspStop <args>`, the current code tries to match a string
with the format "123 (blablabla)", which will result in terminating the
server with ID 123. This is fine when using autocompletion, but if doing
this programmatically (or just following the docs) something like
`:LspStop 1` will not match anything, resulting in terminating *all*
managed servers.

This should be enough to fix the bug, allow terminating multiple servers
at once and keeping the nice UX provided by the autocompletion.