[julials] Let root_dir fallback to the directory of the file
If the file is not in a git repository, and no Project.toml file is
found, this sets root_dir to the directory of the file. Some
functionality in LanguageServer.jl is limited to files inside "workspace
folders"[0]. When rootUri[1] (root_dir) is passed as null to LS.jl this
means that no such folders are identified.

[0]: https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workspaceFolder
[1]: https://microsoft.github.io/language-server-protocol/specifications/specification-current/#initialize