fix(ccls): remove .git from root_dir
CCLS does not use .git to identify the root of a project. In addition,
it causes issues with projects that are based on the repo tool (e.g.
AOSP), which generates .git symlinks in many places in the source tree.
In such cases, CCLS incorrectly identifies subdirectories as the root.