docs(metals): update metals instructions
- Updates link to the metals website
- Removes the instructions about manual bootstrapping

I removed these instructions because with `cs install metals` there is
no reason to manually bootstrap it, we actually discourage it.
Especially now that we are on a new Scala version (2.13), the old
instructions including Scala 2.12 will fail. When using `cs install`,
all of this is just handled for you.

`nvim-metals` is still the path we recommend for any metals users.