fix(lspinfo): safely expand filenames that don't exist on fs
The filename of a buffer may point to a node on the file system that
does not yet exist. `uv.fs_realpath()` will return nil in these cases,
causing a nil exception later on. The builtin `resolve()` function will
return the provided filename as-is if it does not yet exist.

Alternatively we could stick to libuv APIs.