ci: add luarocks upload release workflow
The recently added rockspec (see #2307) has to be prepared and
uploaded to luarocks for each release.

This adds an automated workflow for it, which has been tested with
plenary.nvim.

To be able to upload to luarocks, the owner of the luarocks account
will have to add an API key named `LUAROCKS_API_KEY` to this repo's
GitHub Actions secrets.