feat: add PyrightSetPythonPath command to pyright config
The new command accepts a single path for a python executable that will
be used to reconfigure the pyright language server. This is useful to
change which python will be used by pyright without having to activate
the python environment beforehand. The act of finding which python
executable to use isn't handled here.