fix(zls): use workspace zls.json if available
Language server config is read from `zls.json`. By default common global
and user config directories are searched. However, this omits
per-project configuration, which can be available in the workspace dir.

We already look for `zls.json` to determine root dir, but we don't use
it when starting the language server.

This change appends `cmd` with `--config-path "zls.json"`, which uses
the file when it's available and fall-backs to default logic when file
is not available.