feat(ltex): add plaintext mapping
LTeX will only understand `plaintext`:
https://github.com/valentjn/ltex-ls/blob/16.0.0/src/main/kotlin/org/bsplines/ltexls/parsing/CodeFragmentizer.kt#L70
though it will fall back to plain text when provided `text` as mapping,
but will output a warning when doing so.

Note that we do not enable the `text` filetype by default, as LTeX
itself does not do so. However, this filetype will not be useable
without this mapping anyway, so doing this makes using enabling this
manually slightly easier.