feat(ltex): enable in context, html and xhtml filetypes
ConTeXt was requested in #2663. LTeX is documented to enable it by
default, but this is not consistent with the default in the code:
https://github.com/valentjn/ltex-ls/blob/16.0.0/src/main/kotlin/org/bsplines/ltexls/settings/Settings.kt#L175
Regardless of this we will enable it by ourselves anyway.

We also enable HTML and XHTML as there already has been an attempt to
enable them previously, as there has been a mapping for `xhtml` in the
`language_id_mapping` array. HTML support is enabled in LTeX by default
anyway.