fix(ltex): set ltex.enabled
LTeX is launched whenever the filetype is on the preset list, but will
then verify the language ID against its list and will refuse to run if
the language ID is not on the list. The default can be seen in
https://github.com/valentjn/ltex-ls/blob/16.0.0/src/main/kotlin/org/bsplines/ltexls/settings/Settings.kt#L175
and somewhat matches the supported markup languages list.

The issue resolved by this commit comes from the fact that LTeX will
accept several aliases for each language, for instance it will handle
`gitcommit` and `git-commit` the same way. However, only `git-commit`
appears on the default enabled list. Since there is no `gitcommit ->
git-commit` mapping in `language_id_mapping`, LTeX will ignore
`gitcommit`. This could also be solved by adding this mapping, however
setting a default value of `ltex.enabled` has smaller change of breaking
existing configuration. Same situation happens for `xhtml`.