fix(basepyright): correct command (#3082)
* feat: add basedpyright support

* fix(basedpyright): correct settings schema

Settings are stored under a different key for this fork.  Instead of
using the original `python` key, now settings specific to this LSP
server are stored under `basedpyright` map key.

* fix(basedpyright): correct plugin command

The command was doing nothing because as `python` is no longer the
default namespace for settings, for this LSP server,
`client.settings.python` has a `nil` value.  Hence the table couldn't be
extended.