fix(rescripls): enable typechecking, improve doc (#3281)
I was very confused to find that I got no typechecking by default, since
ReScript is meant to be a very-strongly-typed language. The intent seems
to be to use the "askToStartBuild" option to start an external build,
which integrates with the lsp for typing, but that prompt is much too
buggy, and the incremental typing (enabled in this PR) works well for
me.