feat: show deprecate servers list in LspInfo (#3308)
* feat: show deprecate servers list in LspInfo

In order to smoothly transition to the vim.lsp.start interface in the future,
adding checkhealth here is meaningless. Only the lspinfo information window
is available. So shown deprecate servers list in LspInfo and update the
server deprecate remove version to 0.2.1. 0.2.0 is tagged too
quickly cause some servers not removed.


Co-authored-by: Justin M. Keyes <justinkz@gmail.com>