fix: check buffer validity try_add()/try_add_wrapper()
Currently avoids an issue where a `FileType` autocommand schedules a
call to one of these functions but the buffer gets wiped out before the
next event loop. `nvim_get_option_value()` with the `filetype` argument
can cause this as it creates a transient buffer that only lasts for a
single function call.