fix(fennel-ls): use cwd when git repo parent is not found
Previously when using `fennel-ls`, the LSP would fail to load if no parent was a
Git repository and if a relative `flsproject.fnl` was not found.

Now when using `fennel-ls`, the `root_dir` configuration option is set to the
current working directory as a fallback when all other possible directories
are not available.