fix: remove validation from configs.lua
This is because using the old syntax for vim.validate causes
`:checkhealth vim.deprecated` to be flooded with deprecated messages.

It would also be possible to do a version check and use the newer syntax
for vim.validate, but since configs.lua will be replaced by
vim.lsp.config in the future there is little need to future-proof it.

Closes https://github.com/neovim/nvim-lspconfig/issues/3583.