fix!: remove `lua` as root marker
There is no good reason to have it as a root marker and it makes the
configuration super convoluted. It breaks on even basic setups: opening
`src/nvim/generators/hashy.lua` in the neovim repo causes `src/nvim` to
be the root since there's a `lua` folder there. Instead, just make the
root a flat set of markers and let users who look for a more advanced
setup to adjust it in their own configurations.