fix(bazelrc_lsp): invoke `lsp` subcommand explicitly (#3634)
The bazelrc-lsp binary added other sub-commands besides the `lsp`
functionality (e.g., `bazelrc-lsp format` can be used to format files
from the command line). This commit updates the config such that
the `bazelrc-lsp lsp` is explicitly invoked. (Currently, `bazelrc-lsp`
still defaults to the `lsp` sub-command, but this default will likely be
removed in a future version)