Create vue_ls.lua
Recent changes of renaming of volar-lsp to vue_ls-lsp broke, config resolving on fresh installs. Fix should be, to create a config with name vue_ls.lua.

Btw. The official name of the Vue-LSP is Volar. That's because, there was an much older/obsolete lsp called vetur for vue, that isn't used anymore. In my opinion the recent renaming of lsp and mason packages is confusing. but anyway...


neovim/nvim-lspconfig:
https://github.com/neovim/nvim-lspconfig/pull/3843

mason-registry: 
https://github.com/mason-org/mason-registry/pull/10150