feat(helm_ls,yamlls): support custom filetype for helm-ls #3881
this adds config for starting helm_ls and yamlls with the new filetype
yaml.helm-values that will be added to the neovim helm_ls extension by
https://github.com/qvalentin/helm-ls.nvim/pull/8

Files of this type should have both yamlls and helm-ls running.
This is required because helm-ls now supports editing values.yaml files
but should only be active for certain yaml files.