fix: phpactor root markers precedence
This fixes the precedence of root markers for phpactor. When opening
a vendor file, a new lsp workspace was being created in the context of
the vendor directory, which causes failures because there's no vendor/
directory in the workspace root and therefor the lsp cannot find
any of the symbols.