feat(ts_ls): implement codelens support #3938
Implement the `editor.action.showReferences` client command, which is
used by the references and implementation code lenses that TypeScript
Language Server provides.
I've mostly based the implementation on the existing code for
`vim.lsp.buf.references()`. The main visible difference is that the
latter includes the item being referenced at the top of the list.
Although, if desired, this could be emulated by manually inserting the
position passed along in the command arguments at the top of the list.