feat(gopls): refactor root_dir identification #3942
Refactor the logic on how to identify the root_dir for gopls. This will
provide a viable solution to the issue where multiple gopls lsp client
instances are started, causing adverse effects in certain areas like
diagnostics, lsp functionality (such as `go to definition`,
workspace-enabled projects). The identification for required directories
are based on golang-specific environment variables which are set by
default.