fix(tailwindcss): add fallback for tailwind v4
Since tailwind v4 it is not required to have a tailwind.config.* in
the project 1. The current configuration is preventing the LSP to attach to the configured file types in cases when no tailwind.config.* or none of the other markers (package.json etc.) is present. This commit fixes this limitation by providing .git as a fallback and by documenting this limitation.