fix(docker_language_server): set language id for compose files
This adds a `get_language_id()` handler to the docker language server
config to correctly set the language ID to `dockercompose` as expected
by the server. Before this change, the nvim filetype of
`yaml.docker-compose` was passed to the server as the language ID,
resulting in in a docker compose file being misinterpreted as a
Dockerfile.

See also: e3d837b938bd1d7e84c3eb0362b6761e4d5dc15d / #3982