feat(gh_actions_ls): handle "actions/readFile" request #4046
Fixes https://github.com/lttb/gh-actions-language-server/issues/5

Implement a custom handler for the "actions/readFile" request in the
gh_actions_ls config. This handler reads the requested file from disk and
returns its contents if the file is readable. This improves integration
with the GitHub Actions language server by supporting file content
requests.

Here is how the vscode github actions extension sets the handler:
https://github.com/github/vscode-github-actions/blob/main/src/workflow/languageServer.ts#L68

TODO:
We could also provide an implementation for populating the init_options
properly, docs currently suggest an empty table. For the lsp to work properly,
it should be populated with this "shape":

    init_options = {
      sessionToken = session_token,
      repos = {
        {
          id = 1008200293,
          owner = org_name,
          name = repo_name,
          workspaceUri = "file://" .. vim.fn.getcwd(),
          organizationOwned = true,
        },
      },
    },