fix(angularls): correctly resolve mason Angular package path
Problem:
The current extension path strategy assumes the LS is executed from
within the node_modules dir, which is not the case if it was installed
via mason.

Solution:
Check if the executable path is inside a mason folder, then set the
appropriate path for the mason packages; if not, falls back to the
current strategy.